Renamed program to TextToSpeechDemo from TestToSpeechDemo. Added an asynchronous delay to that demo, instead of a synchronous one.
