+++ George Spyrou - Microsoft Cognitive Services/TextToSpeechDemo/App.config
