+++ George Spyrou - Microsoft Cognitive Services/TextToSpeechDemo/Properties/AssemblyInfo.cs
