+++ George Spyrou - Microsoft Cognitive Services/TextToSpeechDemo/TextToSpeechDemo.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{13FE3E3B-ECAE-445A-8841-DDE3E7A076DB}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>TestToSpeechDemo</RootNamespace>
-    <AssemblyName>TestToSpeechDemo</AssemblyName>
+    <RootNamespace>TextToSpeechDemo</RootNamespace>
+    <AssemblyName>TextToSpeechDemo</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
